requirements.txt
absl-py==2.1.0
astunparse==1.6.3
attrs==24.1.0
bcrypt==4.2.0
blinker==1.8.2
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
dnspython==2.6.1
email-validator==2.2.0
flask==3.0.3
Flask-Admin==1.6.1
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
flask-sqlalchemy==3.1.1
flask-wtf==1.2.1
flatbuffers==24.3.25
fonttools==4.53.1
gast==0.6.0
google-pasta==0.2.0
greenlet==3.0.3
grpcio==1.65.4
h5py==3.11.0
idna==3.7
importlib-metadata==8.2.0
importlib-resources==6.4.0
itsdangerous==2.2.0
jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter-server-terminals==0.5.3
keras==3.4.1
kiwisolver==1.4.5
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.1.post1
mdurl==0.1.2
ml-dtypes==0.4.0
namex==0.0.8
numpy==1.26.4
opencv-python==4.10.0.84
opt-einsum==3.3.0
optree==0.12.1
overrides==7.7.0
packaging==24.1
pandas==2.2.2
pillow==10.4.0
prometheus-client==0.20.0
protobuf==4.25.4
pygame==2.6.0
pygame-ce==2.5.0
pygame-gui==0.6.12
pygame-menu==4.4.3
pygments==2.18.0
pyparsing==3.1.2
pyperclip==1.9.0
python-dateutil==2.9.0.post0
python-i18n==0.3.9
python-json-logger==2.0.7
pytz==2024.1
pywinpty==2.0.13
referencing==0.35.1
requests==2.32.3
rfc3339-validator==0.1.4
rich==13.7.1
rpds-py==0.19.1
scikit-learn==1.5.1
scipy==1.13.1
six==1.16.0
SQLAlchemy==2.0.31
tabulate==0.9.0
tensorboard==2.17.0
tensorboard-data-server==0.7.2
tensorflow==2.17.0
tensorflow-intel==2.17.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
terminado==0.18.1
threadpoolctl==3.5.0
tornado==6.4.1
traitlets==5.14.3
typing-extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
websocket-client==1.8.0
werkzeug==3.0.3
wrapt==1.16.0
wtforms==3.1.2
zipp==3.19.2
